Use MAX Framework package - use the maxfw package - update tests - fix pep issues - add pep testing to travis
