+++ .dockerignore
@@ -1,4 +1,3 @@
-Dockerfile
 README.*
 .idea/
 .git/
